Remove K32EnumProcessModules usage
