+++ src/System.Diagnostics.Process/src/PinvokeAnalyzerExceptionList.analyzerdata.uap
@@ -1,6 +1,6 @@
 advapi32.dll!OpenProcessToken
 kernel32.dll!GetProcessAffinityMask
-kernel32.dll!K32EnumProcessModules
 kernel32.dll!SetProcessAffinityMask
 kernel32.dll!SetThreadAffinityMask
+kernel32.dll!QueryFullProcessImageName
 ntdll.dll!NtQuerySystemInformation
